server/ml_models/requirements.txt
ultralytics==8.3.0
opencv-python==4.10.0.84
numpy==1.26.4
pillow==10.4.0

